refactor(web): split server modules
